PrePreBC
- Lot of stuff with lighting system.
- Destroying tiles underneath the player. Posibility to add attacking to
sides. (Listenerers)
- Can't move to tiles that are too high. Automove when stepped into
blanktile. If there is more blanktiles than fall to the bottom.
- Timer not functional when menu is activated
- Added effects, repaired effects.
- Added Character, Item, Conversation Panel. Character panel shows
effects.
- Effects are saved with save file not entities.
- Effects are updated after each second as the lighting. No not needed
updates.
- Entities are updated only if there are near the player. (possibility
to add more entities - up to now when there would be a lot of entities
than it would be a lot slower)
-Correct saving with added content.
- Player can activate other entity. Action will be executed on
activation (action can be lua or listener ). This approach is used in
showing conversation panel.
- Conversations implemented with everything that should be.